ONCO-56: Moved FhirBundleResultsSetPagination to the separate file
